firmware: fix ~ALERT interrupt handling
The interrupt is level triggered on the FX2, so disable the handler until the
ADC disengages it.

It did somehow "work" without this fix, but the CPU was spending most of it's time
entering the interrupt handler.

Found by mistakenly not properly disengaging the ~ALERT line, then all USB comms
broke. This doesn't happen with this commit.